Backend/requirements.txt
flask
flask-socketio
openai-whisper
yt-dlp
nltk
python-dotenv
requests
